Refactor span processor tests
When using the API we can't rely on attributes to determine the order
of span processors because SetAttributes overwrites attributes when a
for existing keys, which results in keeping only the attribute set by
the last span processor.

Before this change it was possible to use attributes to determine span
processor order because the tests wrote to the attributes slice
directly rather than via the SetAttributes API method.

Use events as a wrapper around attributes instead.